// frontend/src/app/models/global-stats.model.ts
export interface GlobalStats {
  totalUsers: number;
  totalSkins: number;
  totalGames: number;
  coinflipsPlayed: number;
  jackpotsResolved: number;
  battlesWaiting: number;
  marketActive: number;
  marketSold: number;
  totalCoinsWagered: number;
  totalCoinsWon: number;
  totalCoinsDeposited: number;
}